+++ src/modules/types/mod.ts
@@ -74,7 +74,9 @@ export type { ConfigFile } from 'typings/config.ts'
 export type { FormatAvailableFiles, LinterAvailableFiles } from 'typings/files.ts'
 
 export type {
+  ClassFieldDecoratorMeta,
   DefaultTransformValidationOpts,
+  RTOFieldMetadata,
   RtoTypes,
   ValidationConstraints,
   ValidationDecorator,